Windows issue.  Before and after algorithm:
Before:
If there is no file named "asadmin" then this must be Windows, run "asadmin.bat"

After:
If we are running on Windows, look for "asadmin.bat" and run that in preference

I.e. the implicit assumption used to be that Windows installations had no ile named "asadmin" in the bin directory.  That is definitely a wrong
assumption -- at least if you simply unpack the plain zip distribution


svn path=/trunk/; revision=84